requirements.txt
streamlit
pandas
plotly
supabase
python-dotenv
requests